move model/message -> websocket/message
