+++ persistence/model.go
@@ -3,7 +3,7 @@ package persistence
 import (
 	"github.com/sunist-c/genius-invokation-simulator-backend/entity/model"
 	"github.com/sunist-c/genius-invokation-simulator-backend/enum"
-	"github.com/sunist-c/genius-invokation-simulator-backend/model/message"
+	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/websocket/message"
 	"time"
 )
