+++ protocol/http/service/localization.go
@@ -5,8 +5,8 @@ import (
 	"github.com/sunist-c/genius-invokation-simulator-backend/enum"
 	"github.com/sunist-c/genius-invokation-simulator-backend/persistence"
 	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http"
+	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http/message"
 	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http/middleware"
-	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http/model"
 	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http/util"
 )
 
@@ -40,7 +40,7 @@ func queryLanguagePackServiceHandler() func(ctx *gin.Context) {
 		} else if has, record := persistence.LocalizationPersistence.QueryByID(result); !has {
 			ctx.AbortWithStatus(404)
 		} else {
-			response := model.LocalizationQueryResponse{LanguagePack: record.Pack()}
+			response := message.LocalizationQueryResponse{LanguagePack: record.Pack()}
 			ctx.JSON(200, response)
 		}
 	}
@@ -52,7 +52,7 @@ func translateServiceServiceHandler() func(ctx *gin.Context) {
 			exist        bool
 			languagePack string
 			destLanguage int
-			request      model.TranslationRequest
+			request      message.TranslationRequest
 		)
 		if exist, languagePack = util.QueryPath(ctx, "language_package"); !exist {
 			ctx.AbortWithStatus(400)
@@ -64,7 +64,7 @@ func translateServiceServiceHandler() func(ctx *gin.Context) {
 			if has, dictionary := persistence.LocalizationPersistence.QueryByID(languagePack); !has {
 				ctx.AbortWithStatus(404)
 			} else {
-				response := model.TranslationResponse{Translation: map[string]string{}}
+				response := message.TranslationResponse{Translation: map[string]string{}}
 				language := enum.Language(destLanguage)
 				for _, word := range request.Words {
 					if ok, result := dictionary.Translate(word, language); ok {
