+++ protocol/http/service/player.go
@@ -4,8 +4,8 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/sunist-c/genius-invokation-simulator-backend/persistence"
 	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http"
+	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http/message"
 	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http/middleware"
-	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http/model"
 	"github.com/sunist-c/genius-invokation-simulator-backend/protocol/http/util"
 	util2 "github.com/sunist-c/genius-invokation-simulator-backend/util"
 )
@@ -47,7 +47,7 @@ func initPlayerService() {
 
 func loginServiceHandler() func(ctx *gin.Context) {
 	return func(ctx *gin.Context) {
-		request := model.LoginRequest{}
+		request := message.LoginRequest{}
 		if !util.BindJson(ctx, &request) {
 			// RequestBody解析失败，BadRequest
 			ctx.AbortWithStatus(400)
@@ -56,20 +56,20 @@ func loginServiceHandler() func(ctx *gin.Context) {
 			ctx.AbortWithStatus(400)
 		} else if has, player := persistence.PlayerPersistence.QueryByID(uint(id)); !has {
 			// 没找到请求玩家，NotFound，登陆失败
-			ctx.JSON(404, model.LoginResponse{Success: false})
+			ctx.JSON(404, message.LoginResponse{Success: false})
 		} else if success, encodeResult := util2.EncodePassword([]byte(request.Password), uint(id)); !success {
 			// 编码密码失败，InternalError，登陆失败
-			ctx.JSON(500, model.LoginResponse{Success: false})
+			ctx.JSON(500, message.LoginResponse{Success: false})
 		} else if string(encodeResult) != (player.Password) {
 			// 密码校验失败，Forbidden，登陆失败
 			middleware.Interdict(ctx, middlewareConfig)
-			ctx.JSON(403, model.LoginResponse{Success: false})
+			ctx.JSON(403, message.LoginResponse{Success: false})
 		} else if !middleware.AttachToken(ctx, middlewareConfig, uint(id)) {
 			// 生成token失败，InternalError
-			ctx.JSON(500, model.LoginResponse{Success: false})
+			ctx.JSON(500, message.LoginResponse{Success: false})
 		} else {
 			// 登录成功，获取玩家卡组信息后返回登录成功响应
-			response := model.LoginResponse{
+			response := message.LoginResponse{
 				PlayerUID:       player.UID,
 				Success:         true,
 				PlayerNickName:  player.NickName,
@@ -87,7 +87,7 @@ func loginServiceHandler() func(ctx *gin.Context) {
 
 func registerServiceHandler() func(ctx *gin.Context) {
 	return func(ctx *gin.Context) {
-		request := model.RegisterRequest{}
+		request := message.RegisterRequest{}
 		if !util.BindJson(ctx, &request) {
 			// RequestBody解析失败，BadRequest
 			ctx.AbortWithStatus(400)
@@ -112,7 +112,7 @@ func registerServiceHandler() func(ctx *gin.Context) {
 			ctx.AbortWithStatus(500)
 		} else {
 			// 注册Player成功，返回Player信息
-			ctx.JSON(200, model.RegisterResponse{
+			ctx.JSON(200, message.RegisterResponse{
 				PlayerUID:      result.UID,
 				PlayerNickName: request.NickName,
 			})
@@ -122,7 +122,7 @@ func registerServiceHandler() func(ctx *gin.Context) {
 
 func updatePasswordServiceHandler() func(ctx *gin.Context) {
 	return func(ctx *gin.Context) {
-		request := model.UpdatePasswordRequest{}
+		request := message.UpdatePasswordRequest{}
 		if !util.BindJson(ctx, &request) {
 			// RequestBody解析失败，BadRequest
 			ctx.AbortWithStatus(400)
@@ -161,7 +161,7 @@ func updatePasswordServiceHandler() func(ctx *gin.Context) {
 
 func updateNickNameServiceHandler() func(ctx *gin.Context) {
 	return func(ctx *gin.Context) {
-		request := model.UpdateNickNameRequest{}
+		request := message.UpdateNickNameRequest{}
 		if !util.BindJson(ctx, &request) {
 			// RequestBody解析失败，BadRequest
 			ctx.AbortWithStatus(400)
@@ -197,7 +197,7 @@ func updateNickNameServiceHandler() func(ctx *gin.Context) {
 
 func destroyServiceHandler() func(ctx *gin.Context) {
 	return func(ctx *gin.Context) {
-		request := model.DestroyPlayerRequest{}
+		request := message.DestroyPlayerRequest{}
 		if !util.BindJson(ctx, &request) {
 			// RequestBody解析失败，BadRequest
 			ctx.AbortWithStatus(400)
@@ -206,7 +206,7 @@ func destroyServiceHandler() func(ctx *gin.Context) {
 			ctx.AbortWithStatus(400)
 		} else if !request.Confirm {
 			// 确认失败，Success
-			ctx.JSON(200, model.DestroyPlayerResponse{
+			ctx.JSON(200, message.DestroyPlayerResponse{
 				Success: false,
 			})
 		} else if success, player := persistence.PlayerPersistence.QueryByID(uint(id)); !success {
@@ -224,7 +224,7 @@ func destroyServiceHandler() func(ctx *gin.Context) {
 			ctx.AbortWithStatus(500)
 		} else {
 			// 删除成功，Success
-			ctx.JSON(200, model.DestroyPlayerResponse{
+			ctx.JSON(200, message.DestroyPlayerResponse{
 				Success: true,
 			})
 		}
