+++ protocol/websocket/message/initialize.go
